+++ examples/zephyr/nucleo_wba55cg/memfault_demo_app/config/memfault_platform_config.h
@@ -1,0 +1,9 @@
+#pragma once
+
+//! @file
+//!
+//! Copyright (c) Memfault, Inc.
+//! See License.txt for details
+//!
+//! Platform overrides for the default configuration settings in the memfault-firmware-sdk.
+//! Default configuration settings can be found in "memfault/config.h"